+++ test/e2e/btc_staking_pre_approval_e2e_test.go
@@ -247,7 +247,7 @@ func (s *BTCStakingPreApprovalTestSuite) Test2SubmitCovenantSignature() {
 	s.Len(activeDels.Dels, 1)
 
 	activeDel := activeDels.Dels[0]
-	s.True(activeDel.HasCovenantQuorums(s.covenantQuorum))
+	s.True(activeDel.HasCovenantQuorums(s.covenantQuorum, 0))
 }
 
 func (s *BTCStakingPreApprovalTestSuite) Test3SendStakingTransctionInclusionProof() {
@@ -261,7 +261,7 @@ func (s *BTCStakingPreApprovalTestSuite) Test3SendStakingTransctionInclusionProo
 
 	btcDel, err := chain.ParseRespBTCDelToBTCDel(verifiedDelegations[0])
 	s.NoError(err)
-	s.True(btcDel.HasCovenantQuorums(s.covenantQuorum))
+	s.True(btcDel.HasCovenantQuorums(s.covenantQuorum, 0))
 
 	// staking tx hash
 	stakingMsgTx, err := bbn.NewBTCTxFromBytes(btcDel.StakingTx)